refactor: move Tempo client code into client::tempo submodule
Separates payment-agnostic code from Tempo-specific implementations:

src/client/ now contains only agnostic abstractions:
  - PaymentProvider trait, MultiProvider, DynPaymentProvider
  - Fetch extension, PaymentMiddleware, HttpError

src/client/tempo/ contains all Tempo-specific client code:
  - provider.rs (TempoProvider)
  - session_provider.rs (TempoSessionProvider)
  - channel_ops.rs (escrow, vouchers, channel lifecycle)
  - signing.rs (TempoSigningMode, sign_and_encode)
  - tx_builder.rs (TempoTxOptions, gas estimation, credential builder)

Tempo types are still re-exported at client level for backward
compatibility (e.g., mpp::client::TempoProvider still works).
Future payment methods get their own subdirectory.

Amp-Thread-ID: https://ampcode.com/threads/T-019c7408-ac9c-711b-9b10-cedf9f45cca7
