+++ src/client/mod.rs
@@ -21,16 +21,7 @@ mod error;
 mod provider;
 
 #[cfg(feature = "tempo")]
-pub mod channel_ops;
-
-#[cfg(feature = "tempo")]
-mod session_provider;
-
-#[cfg(feature = "tempo")]
-pub mod signing;
-
-#[cfg(feature = "tempo")]
-pub mod tx_builder;
+pub mod tempo;
 
 #[cfg(feature = "client")]
 mod fetch;
@@ -47,14 +38,9 @@ pub use fetch::PaymentExt as Fetch;
 #[cfg(feature = "middleware")]
 pub use middleware::PaymentMiddleware;
 
+// Re-export Tempo types at client level for convenience
 #[cfg(feature = "tempo")]
-pub use provider::TempoProvider;
+pub use tempo::channel_ops::ChannelEntry;
 
 #[cfg(feature = "tempo")]
-pub use channel_ops::ChannelEntry;
-
-#[cfg(feature = "tempo")]
-pub use session_provider::TempoSessionProvider;
-
-#[cfg(feature = "tempo")]
-pub use signing::TempoSigningMode;
+pub use tempo::{TempoProvider, TempoSessionProvider, TempoSigningMode};
