+++ src/client/tempo/mod.rs
@@ -1,0 +1,15 @@
+//! Tempo-specific client implementations.
+//!
+//! Contains the Tempo payment providers, transaction building,
+//! signing strategies, and channel operations.
+
+pub mod channel_ops;
+mod provider;
+mod session_provider;
+pub mod signing;
+pub mod tx_builder;
+
+pub use channel_ops::ChannelEntry;
+pub use provider::TempoProvider;
+pub use session_provider::TempoSessionProvider;
+pub use signing::TempoSigningMode;
